Improve performance of CLI generate.
Doing a `findAll()` and then counting the length of the array is highly inefficient.
Using `findAll()` slowing down the generation of large data sets for testing.
Change the code to an SQL count via the JPA `count()` method.

+++ src/main/java/org/tdl/vireo/cli/Cli.java
@@ -127,14 +127,14 @@ public class Cli implements CommandLineRunner {
 
                         {
                             Random random = new Random();
-                            int idOffset = cliService.countUsers();
+                            long idOffset = cliService.countUsers();
                             User helpfulHarry = null;
 
                             if (expansive) {
                                 helpfulHarry = cliService.createHelpfulHarry(idOffset++, CliService.EMAIL_DATE);
                             }
 
-                            for (int i = itemsGenerated; i < num1 + itemsGenerated; i++) {
+                            for (long i = itemsGenerated; i < num1 + itemsGenerated; i++) {
                                 cliService.operateGenerate(expansive, num2, random, idOffset, helpfulHarry, i);
                                 System.out.print("\r" + (i - itemsGenerated) + " of " + num1 + " generated...");
                             }
